feat: delete user
