+++ locales/settings/en.json
@@ -272,6 +272,10 @@
   "profile.change_password.label": "Change Password",
   "profile.confirm_password.label": "Confirm Password",
   "profile.current_password.label": "Current Password",
+  "profile.delete_account.action": "Delete",
+  "profile.delete_account.email_sent": "An email has been sent to verify the deletion of your account.",
+  "profile.delete_account.label": "Delete Account",
+  "profile.delete_account.no_password": "You need to <Link>set</Link> a password before deleting your account.",
   "profile.email.change": "Change Email",
   "profile.email.changed": "Email changed.",
   "profile.email.changed_verification_sent": "Email to verify the new email has been sent.",
